Moved search series parsing and removed duplicate code
