+++ src/search/search.service.ts
@@ -8,8 +8,7 @@ import {
 import { ItemsPerPage, ResultType } from './entities/search.enum';
 import fetch from 'node-fetch';
 import cheerio from 'cheerio';
-import { SeriesGenre } from 'src/series/entities/type.enum';
-import SharedFunctions from 'src/shared/SharedMethods';
+import { SeriesSearchParser } from './parsers/SeriesSearchParser';
 
 @Injectable()
 export class SearchService {
@@ -60,38 +59,9 @@ export class SearchService {
       throw new Error('Page out of limit');
     }
 
-    const result: SeriesSearchItem[] = [];
-
-    const series = $('.col-12.col-lg-6.p-3.text');
-
-    series.each((i, elem) => {
-      const parsedSeries = new SeriesSearchItem();
-
-      parsedSeries.id = SharedFunctions.getIdfromURL(
-        $(elem).find('.h-100.w-100').attr('href'),
-      );
-
-      parsedSeries.image = $(elem).find('.h-100.w-100>img').attr('src');
-
-      parsedSeries.nsfw = parsedSeries.image ? false : true;
-
-      parsedSeries.title = $(elem).find('u>b').text();
-
-      parsedSeries.description = $(elem).find('.text.flex-grow-1').text();
-
-      parsedSeries.genres = this.parsedGenres(
-        $(elem).find('.textsmall>a').attr('title'),
-      );
-
-      const [year, score] = this.parseYearAndScore(
-        $(elem).find('.text').last().text(),
-      );
-
-      parsedSeries.year = year;
-      parsedSeries.average = score;
-
-      result.push(parsedSeries);
-    });
+    const searchParser = new SeriesSearchParser();
+    await searchParser.parse($);
+    const result = searchParser.getObject();
 
     return [result, totalPages];
   }
@@ -115,38 +85,11 @@ export class SearchService {
     return params;
   }
 
-  private parsedGenres(genres: string): SeriesGenre[] {
-    if (genres.length === 0) {
-      return [];
-    }
-    const splitGenres = genres.split(', ');
-    return splitGenres.map((item) => <SeriesGenre>item);
-  }
-
-  private parseYearAndScore(text: string): [number?, number?] {
-    text = text.replace(/\s+/g, '');
-
-    if (text.includes('-')) {
-      // Year and score
-      const split = text.split('-');
-      const year = Number(split[0]);
-      const score = Number(split[1].split('/')[0]);
-
-      return [year, score];
-    } else {
-      if (text.includes('/')) {
-        // only score
-        return [null, Number(text.split('/')[0])];
-      } else {
-        // only year
-        return [Number(text), null];
-      }
-    }
-
-    return [null, null];
-  }
-
   private getTextInBrackets(string: string): string {
-    return string.match(/\(([^)]+)\)/)[1];
+    const pages = string.match(/\(([^)]+)\)/);
+    if (!pages) {
+      return '1';
+    }
+    return pages[1];
   }
 }
