+++ src/shared/Parser.ts
@@ -1,6 +1,7 @@
+import { CheerioAPI } from 'cheerio';
 import { Page } from 'puppeteer';
 
 export interface Parser<Type> {
-  parse: (page: Page) => Promise<void>;
+  parse: (page: Page | CheerioAPI) => Promise<void>;
   getObject: () => Type;
 }
